feat: ActivitiesControllerにユーザー認証を追加し、権限エラーを処理

+++ packages/server/src/activities/activities.controller.ts
@@ -3,6 +3,7 @@ import {
   ClassSerializerInterceptor,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   Param,
@@ -26,6 +27,8 @@ import { ActivityResponseDto } from './dto/activity.response.dto';
 import { ActivityUpdateDto } from './dto/activity.update.dto';
 
 import { AuthGuard } from '@/auth/auth.guard';
+import { UserResponseDto } from '@/users/dto/user.response.dto';
+import { User } from '@/users/user.decorator';
 
 @UseInterceptors(ClassSerializerInterceptor)
 @SerializeOptions({ type: ActivityResponseDto })
@@ -37,8 +40,11 @@ export class ActivitiesController {
 
   @Post()
   @ApiCreatedResponse({ type: ActivityResponseDto })
-  create(@Body() createActivityDto: ActivityCreateDto) {
-    const userId = 1; // TODO
+  create(
+    @Body() createActivityDto: ActivityCreateDto,
+    @User() user: UserResponseDto,
+  ) {
+    const userId = user.id;
     return this.activitiesService.create({
       data: { ...createActivityDto, user: { connect: { id: userId } } },
     });
@@ -59,17 +65,34 @@ export class ActivitiesController {
 
   @Patch(':id')
   @ApiOkResponse({ type: ActivityResponseDto })
-  update(
+  async update(
     @Param('id') id: number,
     @Body() updateActivityDto: ActivityUpdateDto,
+    @User() user: UserResponseDto,
   ) {
+    const activity = await this.activitiesService.findOneOrThrow({ id });
+    const userId = user.id;
+    // 自分のIDと異なる場合はエラー
+    if (activity.userId !== userId) {
+      throw new ForbiddenException();
+    }
     return this.activitiesService.update({ id, data: updateActivityDto });
   }
 
   @Delete(':id')
   @HttpCode(204)
   @ApiNoContentResponse()
-  async remove(@Param('id') id: number): Promise<void> {
+  async remove(
+    @Param('id') id: number,
+    @User() user: UserResponseDto,
+  ): Promise<void> {
+    const activity = await this.activitiesService.findOneOrThrow({ id });
+    const userId = user.id;
+    // 自分のIDと異なる場合はエラー
+    if (activity.userId !== userId) {
+      throw new ForbiddenException();
+    }
+
     await this.activitiesService.remove({ id });
   }
 }
